Backed out changeset 384e4f88f047 (bug 1784880) for causing wpt failures in parsedepth.html CLOSED TREE

+++ dom/base/nsContentUtils.cpp
@@ -7521,8 +7521,8 @@ Maybe<nsContentUtils::ParsedRange> nsContentUtils::ParseSingleRangeRequest(
     const nsACString& aHeaderValue, bool aAllowWhitespace) {
   // See https://fetch.spec.whatwg.org/#simple-range-header-value
   mozilla::Tokenizer p(aHeaderValue);
-  Maybe<uint64_t> rangeStart;
-  Maybe<uint64_t> rangeEnd;
+  Maybe<uint32_t> rangeStart;
+  Maybe<uint32_t> rangeEnd;
 
   // Step 2 and 3
   if (!p.CheckWord("bytes")) {
@@ -7545,7 +7545,7 @@ Maybe<nsContentUtils::ParsedRange> nsContentUtils::ParseSingleRangeRequest(
   }
 
   // Step 8 and 9
-  uint64_t res;
+  int32_t res;
   if (p.ReadInteger(&res)) {
     rangeStart = Some(res);
   }
